fixed some styling, moved home page to index page.
